Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/registered-property-cssom-expected.txt


PASS CSSOM setters function as expected for unregistered properties 
FAIL CSS.registerProperty The given initial value does not parse for the given syntax.
FAIL Formerly valid values are still readable from inline styles but are computed as the unset value assert_equals: expected "rgb(0, 0, 255)" but got "hello"
FAIL Values not matching the registered type can't be set assert_equals: expected "hello" but got "20"
FAIL Values can be removed from inline styles assert_equals: expected "rgb(255, 0, 0)" but got " red"
PASS Stylesheets can be modified by CSSOM 
FAIL Valid values can be set on inline styles assert_equals: expected "rgb(255, 192, 203)" but got "pink"
